requirements.txt
# Requirements
discord.py
requests
logging